Sparse resources: some code cleanup/minor updates

+++ Graphics/GraphicsEngine/interface/DeviceContext.h
@@ -1683,30 +1683,31 @@ typedef struct UpdateIndirectRTBufferAttribs UpdateIndirectRTBufferAttribs;
 struct SparseBufferMemoryBindRange
 {
     /// Offset in buffer address space where memory will be bound/unbound.
-    /// Must be multiple of the BufferSparseProperties::BlockSize.
+    /// Must be a multiple of the BufferSparseProperties::BlockSize.
     Uint64           BufferOffset  DEFAULT_INITIALIZER(0);
-    
-    /// Memory range offset in the pMemory.
-    /// Must be multiple of the BufferSparseProperties::BlockSize.
-    Uint32           MemoryOffset  DEFAULT_INITIALIZER(0);
-    
+
+    /// Memory range offset in pMemory.
+    /// Must be a multiple of the BufferSparseProperties::BlockSize.
+    Uint64           MemoryOffset  DEFAULT_INITIALIZER(0);
+
     /// Size of the memory which will be bound/unbound.
-    /// Must be multiple of the BufferSparseProperties::BlockSize.
-    Uint32           MemorySize    DEFAULT_INITIALIZER(0);
-    
+    /// Must be a multiple of the BufferSparseProperties::BlockSize.
+    Uint64           MemorySize    DEFAULT_INITIALIZER(0);
+
     /// Pointer to the memory object.
-    /// If non-null then memory will be bound to Region, otherwise memory will be unbound.
-    /// Direct3D11: resource must use a single memory object, when resource bound to a new memory object all previous bindings will be unbound.
-    /// Vulkan: memory object must be compatible with resource, use IDeviceMemory::IsCompatible() to ensure that.
-    /// Vulkan & Direct3D12: resource may have bound memory from different memory objects.
+    /// If non-null, the memory will be bound to Region; otherwise the memory will be unbound.
+    /// Direct3D11: a resource must use a single memory object. When a resource is bound to a new memory object,
+    ///             all previous bindings are invalidated.
+    /// Vulkan & Direct3D12: different resource regions may be bound to different memory objects.
+    /// Vulkan: memory object must be compatible with the resource, use IDeviceMemory::IsCompatible() to ensure that.
     IDeviceMemory*   pMemory       DEFAULT_INITIALIZER(nullptr);
-    
+
 #if DILIGENT_CPP_INTERFACE
     constexpr SparseBufferMemoryBindRange() noexcept {}
 
     constexpr SparseBufferMemoryBindRange(Uint64         _BufferOffset,
-                                          Uint32         _MemoryOffset,
-                                          Uint32         _MemorySize,
+                                          Uint64         _MemoryOffset,
+                                          Uint64         _MemorySize,
                                           IDeviceMemory* _pMemory) noexcept :
         BufferOffset{_BufferOffset},
         MemoryOffset{_MemoryOffset},
@@ -1723,10 +1724,10 @@ struct SparseBufferMemoryBind
     /// Buffer for which sparse binding command will be executed.
     IBuffer*                           pBuffer    DEFAULT_INITIALIZER(nullptr);
 
-    /// Array of memory bind commands.
+    /// An array of NumRanges buffer memory ranges to bins.
     const SparseBufferMemoryBindRange* pRanges    DEFAULT_INITIALIZER(nullptr);
 
-    /// Number of elements in the pRanges.
+    /// The number of elements in pRanges array.
     Uint32                             NumRanges  DEFAULT_INITIALIZER(0);
 };
 typedef struct SparseBufferMemoryBind SparseBufferMemoryBind;
@@ -1734,40 +1735,45 @@ typedef struct SparseBufferMemoryBind SparseBufferMemoryBind;
 /// This structure is used by SparseTextureMemoryBind.
 struct SparseTextureMemoryBindRange
 {
-    /// If equals to TextureSparseProperties::FirstMipInTail then all subsequent mip levels are also affected.
+    /// Mip level to bind.
+    ///
+    /// \note If this level is equal to TextureSparseProperties::FirstMipInTail,
+    ///       all subsequent mip levels will also be affected.
     Uint32           MipLevel      DEFAULT_INITIALIZER(0);
 
     /// Texture array slice index.
     Uint32           ArraySlice    DEFAULT_INITIALIZER(0);
 
     /// Region in pixels where to bind/unbind memory.
-    /// Must be multiple of TextureSparseProperties::TileSize.
-    /// If MipLevel is equals to TextureSparseProperties::FirstMipInTail then this field ignored and OffsetInMipTail must be used.
+    /// Must be a multiple of TextureSparseProperties::TileSize.
+    /// If MipLevel is equal to TextureSparseProperties::FirstMipInTail, this field is ignored
+    /// and OffsetInMipTail is used instead.
     Box              Region        DEFAULT_INITIALIZER({});
-    
+
     /// Used to bind/unbind memory to the mip tail.
-    /// If MipLevel is less than TextureSparseProperties::FirstMipInTail then this field ignored and Region must be used.
+    /// If MipLevel is less than TextureSparseProperties::FirstMipInTail,
+    /// this field is ignored and Region is used.
     Uint64           OffsetInMipTail DEFAULT_INITIALIZER(0);
 
     /// Size of the memory which will be bound/unbound to Region.
-    /// Memory size must equal to the number of tiles in Region multiplied by the sparse block size.
-    /// Must be multiple of the SparseMemoryProperties::SparseBlockSize.
+    /// Memory size must be equal to the number of tiles in Region multiplied by the sparse block size.
+    /// It must be a multiple of the SparseMemoryProperties::SparseBlockSize.
     /// Ignored in Metal.
-    /// 
-    /// \note This is 32bit value because on most devices size of single memory block is limited to 2^32.
-    Uint32           MemorySize    DEFAULT_INITIALIZER(0);
-    
+    Uint64           MemorySize    DEFAULT_INITIALIZER(0);
+
     /// Memory range offset in the pMemory.
-    /// Must be multiple of the TextureSparseProperties::BlockSize.
+    /// Must be a multiple of the TextureSparseProperties::BlockSize.
     /// Ignored in Metal.
-    Uint32           MemoryOffset  DEFAULT_INITIALIZER(0); // AZ TODO: 64bit? In dx11 it is always 32bit, in dx12 & vk can be 64 only because of multiple pages
-    
+    Uint64           MemoryOffset  DEFAULT_INITIALIZER(0);
+
     /// Pointer to the memory object.
-    /// If non-null then memory will be bound to Region, otherwise memory will be unbound.
-    /// Direct3D11: resource must use a single memory object, when resource bound to a new memory object all previous bindings will be unbound.
-    /// Metal: must be the same memory object which is used to create sparse texture, see IRenderDeviceMtl::CreateSparseTexture().
-    /// Vulkan: memory object must be compatible with resource, use IDeviceMemory::IsCompatible() to ensure that.
-    /// Vulkan & Direct3D12: resource may have bound memory from different memory objects.
+    /// If non-null, the memory will be bound to Region; otherwise the memory will be unbound.
+    /// Direct3D11: resource must use a single memory object; when a resource is bound to a new memory
+    ///             object, all previous bindings are invalidated.
+    /// Vulkan & Direct3D12: different resource regions may be bound to different memory objects.    
+    /// Metal: must be the same memory object that was used to create the sparse texture,
+    ///        see IRenderDeviceMtl::CreateSparseTexture().
+    /// Vulkan: memory object must be compatible with the resource, use IDeviceMemory::IsCompatible() to ensure that.
     IDeviceMemory*   pMemory       DEFAULT_INITIALIZER(nullptr);
 };
 typedef struct SparseTextureMemoryBindRange SparseTextureMemoryBindRange;
@@ -1777,11 +1783,11 @@ struct SparseTextureMemoryBind
 {
     /// Texture for which sparse binding command will be executed.
     ITexture*                           pTexture   DEFAULT_INITIALIZER(nullptr);
-    
-    /// Array of memory bind commands.
+
+    /// An array of NumRanges texture memory ranges.
     const SparseTextureMemoryBindRange* pRanges    DEFAULT_INITIALIZER(nullptr);
 
-    /// Number of elements in the pRanges.
+    /// The number of elements in the pRanges array.
     Uint32                              NumRanges  DEFAULT_INITIALIZER(0);
 };
 typedef struct SparseTextureMemoryBind SparseTextureMemoryBind;
@@ -1789,38 +1795,41 @@ typedef struct SparseTextureMemoryBind SparseTextureMemoryBind;
 /// This structure is used by IDeviceContext::BindSparseMemory().
 struct BindSparseMemoryAttribs
 {
-    /// Array of sparse buffer bind commands.
+    /// An array of NumBufferBinds sparse buffer bind commands.
     /// All commands must bind/unbind unique range in the buffer.
     /// Not supported in Metal.
     const SparseBufferMemoryBind*  pBufferBinds   DEFAULT_INITIALIZER(nullptr);
 
-    /// Number of elements in the pBufferBinds.
+    /// The number of elements in the pBufferBinds array.
     Uint32                         NumBufferBinds DEFAULT_INITIALIZER(0);
 
-    /// Array of sparse texture bind commands.
+    /// An array of NumTextureBinds sparse texture bind commands.
     /// All commands must bind/unbind unique region in the texture.
     const SparseTextureMemoryBind* pTextureBinds   DEFAULT_INITIALIZER(nullptr);
 
-    /// Number of elements in the pTextureBinds.
+    /// The number of elements in the pTextureBinds.
     Uint32                         NumTextureBinds DEFAULT_INITIALIZER(0);
-    
-    /// Array of fences to wait.
-    /// Context will wait untill all fences a reached values specified in pWaitFenceValues.
+
+    /// An array of NumWaitFences fences to wait.
+
+    /// \remarks The context will wait untill all fences have reached the values
+    ///          specified in pWaitFenceValues.
     IFence**      ppWaitFences       DEFAULT_INITIALIZER(nullptr);
 
-    /// Array of the values that the context is waiting for the fences to reach.
+    /// An array of NumWaitFences values that the context should wait for the
+    /// fences to reach.
     const Uint64* pWaitFenceValues   DEFAULT_INITIALIZER(nullptr);
 
-    /// Number of elements in the ppWaitFences and pWaitFenceValues.
+    /// The number of elements in the ppWaitFences and pWaitFenceValues arrays.
     Uint32        NumWaitFences      DEFAULT_INITIALIZER(0);
 
-    /// Array of the fences to signal.
+    /// An array of NumSignalFences fences to signal.
     IFence**      ppSignalFences     DEFAULT_INITIALIZER(nullptr);
 
-    /// Array of the values to set the fences to.
+    /// An array of NumSignalFences values to set the fences to.
     const Uint64* pSignalFenceValues DEFAULT_INITIALIZER(nullptr);
-    
-    /// Number of elements in the ppSignalFences and pSignalFenceValues.
+
+    /// The number of elements in the ppSignalFences and pSignalFenceValues arrays.
     Uint32        NumSignalFences    DEFAULT_INITIALIZER(0);
 };
 typedef struct BindSparseMemoryAttribs BindSparseMemoryAttribs;
@@ -3113,12 +3122,12 @@ DILIGENT_BEGIN_INTERFACE(IDeviceContext, IObject)
 
 
     /// Bind or unbind memory objects to sparse buffer and sparse textures.
-    
+
     /// \param [in] Attribs - command attributes, see Diligent::BindSparseMemoryAttribs.
-    /// 
+    ///
     /// \remarks This command will implicitly call Flush().
-    /// 
-    /// \remarks Requires COMMAND_QUEUE_TYPE_SPARSE_BINDING flag in queue.
+    ///
+    /// \remarks Requires COMMAND_QUEUE_TYPE_SPARSE_BINDING flag in the internal queue.
     VIRTUAL void METHOD(BindSparseMemory)(THIS_
                                           const BindSparseMemoryAttribs REF Attribs) PURE;
 };
